Move addClipMarkup method to external functions file Call getVideoObjectMetadata instead of individual getters for clip data
